added menu embr
