+++ src/Components/Navbar/Navbar.module.css
@@ -1,0 +1,99 @@
+
+
+
+
+.NavbarContainer {
+    z-index:999;
+    background-color:#fed2c0 ;
+    padding: 10px ;
+    display: flex;
+    justify-content: space-around;
+    align-items: center;
+    height: 60px;
+    position: relative;
+
+  }
+  
+  .divLogo {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    margin-bottom: 20px;
+    margin-right: 10px; /* Ajusta o empujar a la derecha*/
+    
+  }
+  
+  .imgLogo {
+    width: auto;
+    height: 50px;
+    margin-right: 40px;
+    transition: transform 0.1s;
+  }
+  .divLogo:hover {
+    transform: scale(1.1);
+
+  }
+  .NavbarContainer ul  {
+    display: flex;
+    justify-content: flex-end;
+    list-style: none;
+    padding: 0;
+    margin: 0;
+  }
+
+
+  .NavbarContainer ul li  {
+  margin-right: 10px; /* Ajusta el espaciado entre los enlaces */
+}
+.NavbarContainer ul li a {
+    text-decoration: none;
+    color: black;
+    font-size: 20px;
+    padding: 10px;
+    
+  }
+  .NavbarContainer ul li a:hover {
+    border-bottom: 2px solid red; 
+
+  }
+
+  .mobile-menu-icon{
+    display: none;
+    border: none;
+    background: transparent;
+    font-size: 20px;
+    color: #000;
+    cursor: pointer;
+    margin-left: auto;
+  }
+  @media  screen and(max-width:1999px) {
+    .divLogo{
+      margin-right: auto;
+    } 
+  }
+  .NavbarContainer ul{
+    
+  }
+
+@media (max-width:650px) {
+  .NavbarContainer ul,.open,.close{
+  display: none;
+
+  }
+}
+
+  .mobile-menu-icon {
+   background-color: red;
+  }
+  @media (min-width:650px) {
+    .mobile-menu-icon{
+    display: none;
+    }
+  }  
+  
+
+
+
+
+
+
