+++ src/Components/Navbar/nav.css
@@ -1,0 +1,27 @@
+
+  .mobile-menu-icon {
+    background-color: rgb(239, 232, 232);
+   }
+   @media (min-width:650px) {
+     .mobile-menu-icon{
+     display: none;
+     }
+   }  
+   .open{
+    background-color:#fed2c0;
+    width: 100%;
+position: absolute;
+display: flex;
+gap: 1rem;
+margin-top: 60px;
+top:0;
+flex-direction: column;
+align-items: center;
+font-weight: 700;
+padding: 1rem;
+
+   }
+   .close{
+display: none;
+
+   }
