--- src/components/Navbar/Navbar.module.css
@@ -1,130 +1,0 @@
-
-
-
-
-.NavbarContainer {
-    z-index:999;
-    background-color:#fed2c0 ;
-    padding: 10px ;
-    width: 100%;
-    position: fixed;
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    height: 60px;
-
-  }
-  
-  .divLogo {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    margin-bottom: 20px;
-    margin-right: 40px; /* Ajusta o empujar a la derecha*/
-    
-  }
-  
-  .imgLogo {
-    width: auto;
-    height: 50px;
-    margin-right: 10px;
-    transition: transform 0.1s;
-  }
-  .divLogo:hover {
-    transform: scale(1.1);
-
-  }/*links*/
-  
-input[type="checkbox"] {
-  display: none;
-}
-
-.menuIcon {
-  cursor: pointer;
-}
-
-.menuIcon::before {
-  content: "\2630";
-  font-size: 20px;
-}
-
-
-  nav ul {
-    display:none;
-    list-style-type: none;
-    padding:0;
-    margin:0;
-  }
-  
-  nav ul li {
-    margin-right: 10px;
-  }
-  
-  nav ul li a {
-    color: black;
-    padding: 10px;
-    text-decoration: none;
-    font-size: 1.1em;
-    font-weight: 500;
-    
-  }
-  input[type="checkbox"]:checked ~ nav ul {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-
-  }
-
-/* Media queries */
-@media (min-width: 768px) {
-    .menuIcon {
-     margin-right: 20px;
-    }
-  
-    nav ul {
-      display: flex;
-      justify-content: center;
-    }
-  
-    nav ul li:last-child {
-      margin-right:30px;
-     
-    }
-  }
-
-@media (max-width: 768px) {
-  .divLogo {
-    margin-right: 20px;
-  }
-
-  input[type="checkbox"]:checked ~ nav ul {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-  }
-
-  nav ul li {
-    margin-right: 0;
-    margin-bottom: 10px;
-  }
-}
-
-
-
-
-
-
-  
-
-
-
-
-
-
-
-
-
-
-
-
-
